Stop using GtkWidget event signals: Configure.
Port configure event to default sizes when persisting window state.
This state can also be fetched at an arbitrarty time and doesn't need to
be synchronized with a callback. Solving the year old comment.

This patch does compile with GTK 3 but does not work as intended. GTK 4
is required for this to give the expected results.

https://docs.gtk.org/gtk4/migrating-3to4.html#stop-using-gtkwidget-event-signals
https://developer.gnome.org/documentation/tutorials/save-state.html